requirements.txt
tornado>=5.1
requests>=2.19.1
fuzzywuzzy>=0.17.0
beautifulsoup4>=4.7.1
lxml>=4.3.3
SQLAlchemy>=1.3
authnzerver>=0.0.1.post0.dev58
python-dateutil>=2.8.0
